close the registration for new teams when the last part is verified
